+++ app/src/main/java/top/maary/emojiface/app/EmojiFaceApplication.kt
@@ -1,4 +1,4 @@
-package top.maary.emojiface
+package top.maary.emojiface.app
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
