+++ app/src/main/java/top/maary/emojiface/app/MainActivity.kt
@@ -1,11 +1,11 @@
-package top.maary.emojiface
+package top.maary.emojiface.app
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import dagger.hilt.android.AndroidEntryPoint
-import top.maary.emojiface.ui.EditScreen
+import top.maary.emojiface.ui.edit.EditScreen
 import top.maary.emojiface.ui.theme.EmojiFaceTheme
 
 @AndroidEntryPoint
