+++ app/src/main/java/top/maary/emojiface/ui/edit/model/EmojiDetection.kt
@@ -1,0 +1,9 @@
+package top.maary.emojiface.ui.edit.model
+
+data class EmojiDetection(
+    val xCenter: Float,
+    val yCenter: Float,
+    val diameter: Float,
+    val angle: Float,
+    var emoji: String
+)
