+++ app/src/main/java/top/maary/emojiface/ui/edit/model/EmojiList.kt
@@ -1,4 +1,4 @@
-package top.maary.emojiface
+package top.maary.emojiface.ui.edit.model
 
 import kotlinx.serialization.Serializable
