+++ app/src/main/java/top/maary/emojiface/ui/edit/state/ShareEvent.kt
@@ -1,0 +1,8 @@
+package top.maary.emojiface.ui.edit.state
+
+import android.net.Uri
+
+sealed class ShareEvent {
+    data class ShareImage(val uri: Uri) : ShareEvent()
+    data class Error(val message: String) : ShareEvent()
+}
