+++ app/src/main/java/top/maary/emojiface/util/AndroidExtensions.kt
@@ -1,0 +1,14 @@
+package top.maary.emojiface.util
+
+import android.content.Intent
+import android.os.Build
+import android.os.Parcelable
+
+@Suppress("DEPRECATION")
+inline fun <reified T : Parcelable> Intent.getParcelableExtraCompat(name: String): T? {
+    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+        getParcelableExtra(name, T::class.java)
+    } else {
+        getParcelableExtra(name) as? T
+    }
+}
