+++ app/src/main/java/top/maary/emojiface/util/Constants.kt
@@ -1,4 +1,4 @@
-package top.maary.emojiface
+package top.maary.emojiface.util
 
 object Constants {
     const val DEFAULT_FONT_MARKER = "DEFAULT"
